pyproject.toml
[project]
name = "equity-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=1.2.0",
    "crewai-tools>=1.2.0",
    "polars>=1.34.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "treelib>=1.8.0",
]
